% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsim.r
\name{add_sim_bound}
\alias{add_sim_bound}
\title{function dedicated to add simulation standard error bound, in development
draw the bound to current plot}
\usage{
add_sim_bound(
  data,
  family = gaussian(),
  M = 200,
  n = 1000,
  true.theta = c(1, 1, 1)/sqrt(3)
)
}
\arguments{
\item{data}{a list of simulated data}

\item{family}{default is gaussian()}

\item{M}{number of simulations}

\item{n}{sample size}

\item{true.theta}{the true coefficients}
}
\description{
function dedicated to add simulation standard error bound, in development
draw the bound to current plot
}
