% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsim.r
\name{generate_data}
\alias{generate_data}
\title{Data generation function for simulation and demonstration
A sine-bump setting has been employed.}
\usage{
generate_data(
  n,
  true.theta = c(1, 1, 1)/sqrt(3),
  family = "gaussian",
  ncopy = 1
)
}
\arguments{
\item{n}{sample size}

\item{true.theta}{true single-index coefficients,
default is c(1,1,1)/sqrt(3) for setting 1 and c(1,2)/sqrt(5) for other settings}

\item{family}{chose from "gaussian", "binomial" or "poisson".}

\item{ncopy}{generates multiple copies of data for Monte Carlo simulations}
}
\value{
X single index predictors

Y response variables, a list

Z partial linear predictor(s)

single_index_values single index term
}
\description{
Data generation function for simulation and demonstration
A sine-bump setting has been employed.
}
