% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsim.r
\name{plot.si}
\alias{plot.si}
\title{Function that plot fitted curve for the unknown univariate function for single index term}
\usage{
\method{plot}{si}(x, ...)
}
\arguments{
\item{x}{the gam/gplism fitted object}

\item{...}{add-on arguments}
}
\value{
NULL single-index plot
}
\description{
Function that plot fitted curve for the unknown univariate function for single index term
}
