% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot_si}
\alias{plot_si}
\title{Function that plot fitted curve for the unknown univariate function for single index term}
\usage{
plot_si(
  x,
  family = gaussian(),
  ylab = "mean",
  yscale = NULL,
  plot_data = FALSE
)
}
\arguments{
\item{x}{the gam/gplism fitted object}

\item{family}{default is gaussian()}

\item{ylab}{y label}

\item{yscale}{scale of y}

\item{plot_data}{controls whether to plot the data as points}
}
\value{
NULL single-index plot
}
\description{
Function that plot fitted curve for the unknown univariate function for single index term
}
