% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Predict.matrix.tr.smooth}
\alias{Predict.matrix.tr.smooth}
\title{prediction method function for the tr smooth class}
\usage{
Predict.matrix.tr.smooth(object, data)
}
\arguments{
\item{object}{smooth object for gam class}

\item{data}{the new data to predict on
'}
}
\value{
X the prediction matrix
}
\description{
prediction method function for the tr smooth class
}
