% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.summary.gplsim}
\alias{print.summary.gplsim}
\title{Print Summary function of gplsim object}
\usage{
\method{print}{summary.gplsim}(
  x,
  digits = max(5, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{the gam/gplism fitted object}

\item{digits}{controls number of digits printed in output.}

\item{signif.stars}{should significance stars be printed alongside output.}

\item{...}{optional arguments}
}
\value{
summarized object with nice format
}
\description{
Print Summary function of gplsim object
}
