% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsim.r
\name{smooth.construct.tr.smooth.spec}
\alias{smooth.construct.tr.smooth.spec}
\title{supporting function to make tr smooth}
\usage{
smooth.construct.tr.smooth.spec(object, data, knots)
}
\arguments{
\item{object}{smooth object for gam class}

\item{data}{the new data to predict on}

\item{knots}{knots}
}
\value{
tr smooth object
}
\description{
supporting function to make tr smooth
}
