% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsim.r
\name{summary.gplsim}
\alias{summary.gplsim}
\title{Summary function of gplsim object}
\usage{
\method{summary}{gplsim}(object, ...)
}
\arguments{
\item{object}{the gam/gplism fitted object}

\item{...}{optional arguments}
}
\value{
gplsim_obj a list of summary information for a fitted gplsim object, which extends on gam object.
}
\description{
Summary function of gplsim object
}
