% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsim.r
\name{si}
\alias{si}
\title{An internal function to optimization and fitting.
Don't use it solely.}
\usage{
si(
  alpha,
  y,
  x,
  z,
  opt = TRUE,
  smooth_selection,
  fam,
  bs = "ps",
  fx = FALSE,
  scale = scale,
  ...
)
}
\arguments{
\item{alpha}{single-index coefficients}

\item{y}{Response variable, should be a vector.}

\item{x}{Single index covariates.}

\item{z}{Partially linear covariates.}

\item{opt}{see ?gplsim}

\item{smooth_selection}{see ?gplsim}

\item{fam}{see ?gplsim}

\item{bs}{see ?gplsim}

\item{fx}{see ?gplsim}

\item{scale}{see ?gplsim}

\item{...}{includes optional arguments user can pass to \code{mgcv::gam} or \code{glm}, such as \code{k}, which is the dimension of the basis of the smooth term and \code{m}, which is the order of the penalty for the smooth term}
}
\value{
b fitted gam object
}
\description{
An internal function to optimization and fitting.
Don't use it solely.
}
